Modificación UserDashboard en componentes
